moved files around. rviz plugins is still not compiling
