+++ igvc_platform/CMakeLists.txt
@@ -177,4 +177,4 @@ endif()
 add_subdirectory(src/imu)
 add_subdirectory(src/joystick_driver)
 add_subdirectory(src/motor_controller)
-add_subdirectory(src/system_stats)
+#add_subdirectory(src/system_stats)
