+++ igvc_rviz_plugins/src/rviz_plugins/CMakeLists.txt
@@ -16,7 +16,7 @@ add_definitions(-DQT_NO_KEYWORDS)
 
 qt5_wrap_cpp(MOC_FILES
   bat_panel.h
-	eStop_panel.h
+  eStop_panel.h
   example_panel.h
   launch_panel.h
   my_thread.h
@@ -24,14 +24,14 @@ qt5_wrap_cpp(MOC_FILES
   num_button.h
   sensor_panel.h
   speedometer.h
-	speed_panel.h
-	time_panel.h
+  speed_panel.h
+  time_panel.h
   waypoint_panel.h
 )
 
 set(SOURCE_FILES
   bat_panel.cpp
-	eStop_panel.cpp
+  eStop_panel.cpp
   example_panel.cpp
   launch_panel.cpp
   my_thread.cpp
@@ -40,9 +40,9 @@ set(SOURCE_FILES
   sensor_panel.cpp
   speedometer.cpp
   speed_panel.cpp
-	time_panel.cpp
+  time_panel.cpp
   waypoint_panel.cpp
-	${MOC_FILES})
+  ${MOC_FILES})
 
 add_library(${PROJECT_NAME} ${SOURCE_FILES})
 add_dependencies(${PROJECT_NAME} igvc_msgs_gencpp)
