+++ igvc_utils/include/igvc_utils/StringUtils.hpp
